Into the Gray Zone The Private Sector and Active Defense Against Cyber Threats
OVER THE PAST SEVERAL DECADES, the private sector in the United States has em- braced the computer revolution and the growth of the Internet, and migrated its business activities and operations into an information technology environment. This transition to the online domain has provided tremendous benefits to the private sector, enabling business efficiencies, lowering transaction costs, establishing new products and markets, enhancing internal collaboration, and improving the ability of companies to measure and assess their performance. But as the online domain has developed over the past several decades, new risks have accompanied these benefits; companies have become increasingly vulnerable to the theft of online intellectual property or customer data and the disruption of business operations.
These cyber risks and dependencies have grown in recent years due to the activities of hostile state and non-state actors in cyberspace, who have attacked private sector entities for both political and economic reasons. Companies have enhanced their defenses, and the federal government has placed a higher priority on assisting the private sector, but such measures are not commensurate with the nature of the cyber threat today.
This paper examines a set of capabilities that can help to address this gap, collectively defined under the term active defense:
Active defense is a term that captures a spectrum of proactive cybersecurity measures that fall between traditional passive defense and offense. These activities fall into two general categories, the first covering technical interactions between a defender and an attacker. The second category of active defense includes those operations that enable defenders to collect intelligence on threat actors and indicators on the Inter- net, as well as other policy tools (e.g. sanctions, indictments, trade remedies) that can modify the behavior of malicious actors. The term active defense is not synonymous with “hacking back” and the two should not be used interchangeably.