Hurricane Matthew Electric Grid Impact
The U.S. Department of Homeland Security (DHS)/Office of Cyber and Infrastructure Analysis (OCIA) produces Infrastructure Impact Assessments to provide an overview of risks to critical infrastructure from all hazards. The information in this assessment is intended to inform DHS leadership and partners on the current conditions and most critical concerns related to Hurricane Matthew, which is forecast to approach Florida by Friday, October 7, 2016. This product describes the potential effects in Florida and Georgia to the Energy Sector – Electricity Subsector from Hurricane Matthew. The information contained in this assessment is based on National Oceanic and Atmospheric Administration forecasts, open source reporting, and OCIA modeling. This product has been developed using the 0800 EDT, October 6, 2016, National Hurricane Center forecast.