Joint Operating Environment 2035
Joint Operating Environment 2035 -The Joint Force in a Contested and Disordered World
The Joint Operating Environment 2035 (JOE 2035) is designed to encourage the purposeful preparation of the Joint Force to effectively protect the United States, its interests, and its allies in 2035. For the Joint Force, thinking through the most important conditions in a changing world can mean the difference between victory and defeat, success and failure, and the needless expenditure of human lives and national treasure versus the judicious and prudent application of both to defend our vital interests.
This document describes the future security environment and projects the implications of change for the Joint Force so it can anticipate and prepare for potential conflicts. To do this, Section 1 describes the circumstances that are likely to alter the security environment. Next, Section 2 explores how the intersection and interaction of these changes might impact the character of war in the future. Finally, Section 3 provides a framework to think about the full range of Joint Force missions and how they may evolve over time.
JOE 2035 illustrates several ideas about how changes to conflict and war might impact the capabilities and operational approaches required by the future Joint Force.