DHS Report on Cyber Attacks Against Massachusetts
This document is intended to provide network security managers, information security officers and those responsible for cyber security in their respective organizations potentially malicious Internet Protocol addresses in Massachusetts and nationwide. These hosts have been associated with malicious code incidents. (Attached find an excel spreadsheet labelled “Massachusetts Malicious IPs and Domains November 2015” for the full list.)This is being provided for informational purposes only and not to be used in the course of a criminal investigation.