U.S. National Space Weather Strategy
Space weather refers to variations in the space environment between the sun and Earth (and throughout the solar system) that can affect technologies in space and on Earth. Space weather can disrupt the technology that forms the backbone of this country’s economic vitality and national security, including satellite and airline operations, communications networks, navigation systems, and the electric power grid. As the Nation becomes ever more dependent on these technologies, space weather poses an increasing risk to infrastructure and the economy. Further, the Strategic National Risk Assessment1 has identified space weather as a hazard that poses significant risk to the security of the Nation. Clearly, reducing vulnerability to space weather needs to be a national priority.
The National Space Weather Strategy (Strategy) and the accompanying National Space Weather Action Plan (Action Plan) together seek to enhance the integration of existing national efforts and to add important capabilities to help meet growing demands for space-weather information. The Strategy and Action Plan build on recent efforts to reduce risks associated with natural hazards and improve resilience of essential facilities and systems,2 aiming to foster a collaborative environment in which government, industry, and the American people can better understand and prepare for the effects of space weather. The Nation must continue to leverage existing public and private networks of expertise and capabilities and pursue targeted enhancements to improve the ability to manage risks associated with space weather.