National Strategy for Network of Fusion Centers
The terrorist attacks of September 11, 2001, galvanized law enforcement leaders across the nation to improve the sharing of criminal intelligence needed to prevent future terrorist attacks. The resulting rapid evolution of fusion centers has underscored the need to formalize a forward-looking, national-level strategy speci c to the fusion center network. The 2014–2017 National Strategy for the National Network of Fusion Centers (2014–2017 National Strategy) establishes a vision, a mission, goals, objectives, and initiatives that are needed for the National Network of Fusion Centers (NNFC or National Network) to systematically improve intelligence information sharing beyond existing and successful criminal intelligence in support of law enforcement investigations.