12 Risks That Threaten Human Civilisation
This report has, to the best of the authors’ knowledge, created the  first list of global risks with impacts that for all practical purposes can be called in nite. It is also the  first structured overview of key events related to such challenges and has tried to provide initial rough quantifications for the probabilities of these impacts.