Outplayed - Regaining Strategic Initiative in the Gray Zone
U.S. competitors pursuing meaningful revision or rejection of the current U.S.-led status quo are employing a host of hybrid methods to advance and secure interests that are in many cases contrary to those of the United States. These challengers employ unique combinations of in uence, intimidation, coercion, and aggression to incremen- tally crowd out effective resistance, establish local or regional advantages, and manipu- late risk perceptions in their favor.
So far, the United States has not come up with a coherent countervailing approach. It is in this “gray zone”—the awkward and uncomfortable space between traditional conceptions of war and peace—where the United States and its defense enterprise face systemic challenges to U.S. position and authority. As a result, gray zone competition and con ict should be pacers for defense strategy.