Special Forces Guide to Information Operations
IO should be viewed as an element of combat power, focused when and where it best supports the operation. As with other elements of combat power, there is no universal formula for the application of IO. Mission, enemy, terrain and weather, troops and support available-time available, and civil considerations are the major determinants.
The purpose of IO is to achieve and maintain information superiority or advantage over the adversary at a particular time and place. To achieve an information advantage, an SF unit must understand the characteristics of the information environment in its operational area. The unit must also understand how adversary and third-party organizations use information to achieve their objectives.