Healthcare and Public Health Sector Cyberdependencies
Healthcare and Public Health Sector Cyberdependencies
Prepared By: The U.S. Department of Homeland (DHS)/Office of Cyber and Infrastructure Analysis and the DHS/Office of Intelligence and Analysis.
The Department of Homeland Security (DHS) assesses that given the high value of patient information and proprietary data on the black market, the Healthcare and Public Health Sector will continue to be one of the primary targets for malicious cyber actors. Stolen health data sells on the black market for more than 10 to 20 times the price of stolen credit card data. DHS assesses that growth in the medical device market over the next 4 years will result in more devices connected to the Internet, and an increase in the number of cyber-related incidents that target those devices. This is partly because manufacturers do not place enough emphasis on the security of medical devices.