Russian New Generation Warfare Handbook
Since the collapse of the Soviet Union, a number of conflicts and reforms have shaped the Russian military into what it is today. Russia has observed the American lessons learned in Iraq and Afghanistan, as well as their own from the 2008 invasion of Georgia, and applied these to the development of their forces. Several key changes have been implemented under Vladimir Putin (the current President of the Russian Federation) in the previous decade. These include a rapid modernization and increased complexity of equipment, attempts to professionalize military education, and a restructuring of the military to a more expeditionary model in order to project Russian influence into Europe, the Middle East, and Asia.